Rename components to be more semantic
